+++ src/vs/workbench/contrib/chat/browser/actions/chatAttachPromptAction/dialogs/askToSelectPrompt/utils/createPlaceholderText.ts
@@ -1,0 +1,45 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { localize } from '../../../../../../../../../nls.js';
+import { ALT_KEY_NAME, SUPER_KEY_NAME } from '../constants.js';
+import { ISelectPromptOptions } from '../askToSelectPrompt.js';
+
+/**
+ * Creates a placeholder text to show in the prompt selection dialog.
+ */
+export const createPlaceholderText = (options: ISelectPromptOptions): string => {
+	const { widget } = options;
+
+	let text = localize(
+		'commands.prompts.use.select-dialog.placeholder',
+		'Select a prompt to use',
+	);
+
+	// if no widget reference is provided, add the note about `options`
+	// and `cmd` modifiers users can use to alter the command behavior
+	if (!widget) {
+		const altOptionModifierNote = localize(
+			'commands.prompts.use.select-dialog.alt-modifier-note',
+			'{0}-key to use in Copilot Edits',
+			ALT_KEY_NAME,
+		);
+
+		const superModifierNote = localize(
+			'commands.prompts.use.select-dialog.super-modifier-note',
+			'{0}-key to edit',
+			SUPER_KEY_NAME,
+		);
+
+		text += localize(
+			'commands.prompts.use.select-dialog.modifier-notes',
+			' (hold {0} or {1})',
+			altOptionModifierNote,
+			superModifierNote,
+		);
+	}
+
+	return text;
+};
